requirements.txt
coverage==7.2.7
django==4.2.3
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-filter==23.2
django-model-utils==4.3.1
django-typed-models==0.13.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
mesh==1.0.0a1
pdoc3==0.10.0
pre-commit==3.3.3
prospector==1.10.2
pylint-django==2.5.3
pylint==2.17.4
pylint-plugin-utils==0.7
python-utils==3.7.0
pyyaml==6.0